--- a/lib/Dialect/Arc/ArcOps.cpp
+++ b/lib/Dialect/Arc/ArcOps.cpp
@@ -114,6 +114,17 @@ LogicalResult DefineOp::verifyRegions() {
   return success();
 }
 
+bool DefineOp::isPassthrough() {
+  if (getNumArguments() != getNumResults())
+    return false;
+
+  return llvm::all_of(
+      llvm::zip(getArguments(), getBodyBlock().getTerminator()->getOperands()),
+      [](const auto &argAndRes) {
+        return std::get<0>(argAndRes) == std::get<1>(argAndRes);
+      });
+}
+
 //===----------------------------------------------------------------------===//
 // OutputOp
 //===----------------------------------------------------------------------===//