--- a/lib/Dialect/Arc/Transforms/CMakeLists.txt
+++ b/lib/Dialect/Arc/Transforms/CMakeLists.txt
@@ -6,6 +6,7 @@ add_circt_dialect_library(CIRCTArcTransforms
   InferStateProperties.cpp
   InlineArcs.cpp
   InlineModules.cpp
+  LatencyRetiming.cpp
   LegalizeStateUpdate.cpp
   LowerClocksToFuncs.cpp
   LowerLUT.cpp