Currently, the latency attribute on the `arc.state` operations is not really used except for differentiating between 0 and 1 latency. This pass collapses sequences of `arc.state` operations that form a shift register into one state operation. It also aggregates the latencies of isolated fan-in trees. It's still very simple in its current form, but can be easily extended in the future. Since we don't have a proper benchmark running yet to assess performance, I gathered some statistics instead by ignoring the name attributes on the states (basically assuming that the user tells arcilator that they are only interested in the inputs and outputs, but not observing internal state and thus no taps are inserted). I don't add it to arcilator for now since we don't have a way to remove the name attributes in a controlled manner yet (it's also possible to apply these patterns in the presence of taps, but that's also not implemented yet and too much to add to this PR) and also no lowering support for states with >1 latency.

LargeBoomConfig statistics:
```
LatencyRetiming
  (S)  93 latency-units-saved - Number of latency units saved by merging them in a successor state
  (S) 245 num-ops-removed     - Number of zero-latency passthrough states removed
```

| latency | number of `arc.state` before | after |
|---|---|---|
| 0 | 25441 | 25572 |
| 1 | 27175 | 26596 |
| 2 | 0 | 123 |
| 3 | 0 | 80 |

Rocket statistics:
```
LatencyRetiming
  (S) 0 latency-units-saved - Number of latency units saved by merging them in a successor state
  (S) 5 num-ops-removed     - Number of zero-latency passthrough states removed
```